+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 0.1.2 - tbd
+## Version 0.1.2 - 08.01.24
 
-### Added
+### Fixed
+
+- Set websocket default kind to `ws`
+- Mock request response more complete
+- More robust setup and error handling
+- Redis off per default
+- Refactoring
 
 ## Version 0.1.1 - 22.12.23